requirements.txt
lmdb==1.4.1
tqdm==4.66.1
datasets
tensorboardX
accelerate
transformers==4.40.1
pycocotools==2.0.7
scikit-image
Pillow==9.5.0
opencv-python-headless
datasets
einops==0.8.0
transformers[sentencepiece]
sentencepiece
xformers==0.0.23.0
pydantic==1.9.0
deepspeed
peft==0.11.1
diffusers==0.33.1
rotary-embedding-torch==0.6.4
tiktoken==0.7.0
transformers_stream_generator==0.0.5
lmdeploy
flash_attn
timm
OpenEXR